Fix 64-bit pointer truncation: low-address pool + DMA alignment bypass
Two root causes of the SIGILL crash on Linux:

1. RomCopy alignment checks silently skip DMA on 64-bit:
   The N64 DMA controller required 8-byte destination alignment.
   RomCopy checks `(u32)dest & 7` and skips the transfer on failure.
   On Linux, the heap returned addresses above 0x7F0000000000 that
   appeared misaligned after u32 truncation, causing DMA to be silently
   skipped. Compressed data was never copied, decompression ran on
   uninitialised memory, producing garbage → SIGILL.
   Fix: bypass alignment checks on Linux (memcpy has no alignment
   requirements).

2. Pervasive (int)/(s32)/(u32) pointer casts truncate 64-bit addresses:
   100+ locations in the decompiled N64 code do pointer arithmetic via
   `(void *)((s32)ptr + offset)` or `(int)ptr`. With system malloc
   returning high addresses, these casts lose the upper 32 bits.
   Fix: replace the malloc-based heap shim with a pool allocator backed
   by mmap(MAP_32BIT), which guarantees all addresses are in the lower
   2 GB. This makes (s32)/(u32) pointer casts lossless without needing
   to patch every cast site.

Also fix (s32) casts on BorgListingPointer/borgFilesPointer to use
uintptr_t for correctness even with the low-address pool.

+++ src_pseudo/n64Borg/borgmain.cpp
@@ -75,7 +75,7 @@ s16 get_borg_listing_type(s32 param_1){
     listing.Type = -1;
   }
   else {
-    ROMCOPYS(&listing,(void *)((s32)BorgListingPointer + param_1 * 0x10 + 8),0x10,0x1c6);
+    ROMCOPYS(&listing,(void *)((uintptr_t)BorgListingPointer + param_1 * 0x10 + 8),0x10,0x1c6);
     swapBorgListing(&listing);
   }
   return listing.Type;
@@ -94,7 +94,7 @@ s16 GetBorgItemInfo(BorgListing *itemInfo,s32 param_2){ //orphaned, low priority
     sVar1 = -1;
   }
   else {
-    ROMCOPYS(itemInfo,(void *)((s32)BorgListingPointer + param_2 * 0x10 + 8),0x10,488);
+    ROMCOPYS(itemInfo,(void *)((uintptr_t)BorgListingPointer + param_2 * 0x10 + 8),0x10,488);
     swapBorgListing(itemInfo);
     sVar1 = itemInfo->Type;
   }
@@ -120,14 +120,14 @@ borgHeader * getBorgItem(s32 index){
     CRASH("n64Borg.cpp, GetBorgItem()",errmsg);
   }
   else{
-    ROMCOPYS(&listing,(void *)((s32)BorgListingPointer + index * sizeof(BorgListing) + 8),sizeof(BorgListing),541);
+    ROMCOPYS(&listing,(void *)((uintptr_t)BorgListingPointer + index * sizeof(BorgListing) + 8),sizeof(BorgListing),541);
     swapBorgListing(&listing);
     if ((((listing.Type < 3) || (listing.Type == 6)) || (listing.Type == 11)) || (((listing.Type == 12 || (listing.Type == 13)) || (listing.Type == 14)))) {
       if (borgFlag == 0) {
         ALLOCS(ret,gBorgHeaderSizes[listing.Type] + sizeof(void*),561);
         if (gBorgBytes[index] == 0) {
           ALLOCS(borgfile,listing.uncompressed,566);
-          decompressBorg((void *)((s32)borgFilesPointer + listing.Offset),listing.compressed,
+          decompressBorg((void *)((uintptr_t)borgFilesPointer + listing.Offset),listing.compressed,
                          borgfile,listing.uncompressed,(s32)listing.Compression);
           (*borg_funcs_a[listing.Type])(borgfile);
           gBorgpointers[index] = (borgHeader*)borgfile;
@@ -143,7 +143,7 @@ borgHeader * getBorgItem(s32 index){
       else {
         ALLOCS(ret,gBorgHeaderSizes[listing.Type] + sizeof(void*),600);
         ALLOCS(borgfile,listing.uncompressed,602);
-        decompressBorg((void *)((s32)borgFilesPointer + listing.Offset),listing.compressed,borgfile,
+        decompressBorg((void *)((uintptr_t)borgFilesPointer + listing.Offset),listing.compressed,borgfile,
                        listing.uncompressed,(s32)listing.Compression);
         (*borg_funcs_a[listing.Type])(borgfile);
         gBorgpointers[index] = NULL;
@@ -158,7 +158,7 @@ borgHeader * getBorgItem(s32 index){
         size = gBorgHeaderSizes[type] + listing.uncompressed;
         ALLOCS(ret,size,627);
         bzero(ret,size);
-        decompressBorg((void *)((s32)borgFilesPointer + listing.Offset),listing.compressed,
+        decompressBorg((void *)((uintptr_t)borgFilesPointer + listing.Offset),listing.compressed,
                        (u8 *)((s32)ret + gBorgHeaderSizes[listing.Type]),listing.uncompressed,
                        (s32)listing.Compression);
         (*borg_funcs_a[listing.Type])(ret);
@@ -173,7 +173,7 @@ borgHeader * getBorgItem(s32 index){
           size = gBorgHeaderSizes[type] + listing.uncompressed;
           ALLOCS(ret,size,653);
           bzero(ret,size);
-          decompressBorg((void *)((s32)borgFilesPointer + listing.Offset),listing.compressed,
+          decompressBorg((void *)((uintptr_t)borgFilesPointer + listing.Offset),listing.compressed,
                          (u8 *)((s32)ret + gBorgHeaderSizes[listing.Type]),listing.uncompressed,
                          (s32)listing.Compression);
           (*borg_funcs_a[listing.Type])(ret);
